finalizando o projeto

+++ app.py
@@ -6,7 +6,7 @@ import bcrypt
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = "your_secret_key"
-app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://admin:admin123@127.0.0.1:3306/flask-crud'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql:///root:admin123@127.0.0.1:3306/flask-crud'
 
 login_manager = LoginManager()
 db.init_app(app)
@@ -14,6 +14,8 @@ login_manager.init_app(app)
 #view login
 login_manager.login_view = 'login'
 #Session <- conexão ativa
+
+@login_manager.user_loader
 def load_user(user_id):
   return User.query.get(user_id)
 
@@ -22,13 +24,16 @@ def login():
   data = request.json
   username = data.get("username")
   password = data.get("password")
+
   if username and password:
    # Login
     user = User.query.filter_by(username=username).first()
+
     if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
       login_user(user)
       print(current_user.is_authenticated)
       return jsonify({"message": "Autenticação realizada com sucesso!"})
+
   return jsonify({"message": "Credenciais inválidas"}), 400
 
 @app.route('/logout', methods=['GET'])
@@ -42,18 +47,21 @@ def create_user():
   data = request.json
   username = data.get("username")
   password = data.get("password")
+
   if username and password:
     hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
-    user = User(username=username, password=hashed_password, role='user')    
+    user = User(username=username, password=hashed_password, role='user')
     db.session.add(user)
     db.session.commit()
     return jsonify({"message": "Usuario cadastrado com sucesso"})
+
   return jsonify({"message": "Dados invalidos"}), 400
 
 @app.route('/user/<int:id_user>', methods=["GET"])
 @login_required
 def read_user(id_user):
   user = User.query.get(id_user)
+
   if user:
     return {"username": user.username}
   
@@ -64,9 +72,14 @@ def read_user(id_user):
 def update_user(id_user):
   data = request.json
   user = User.query.get(id_user)
+
+  if id_user != current_user.id and current_user.role == "user":
+    return jsonify({"message": "Operação não permitida"}), 403
+
   if user and data.get("password"):
     user.password = data.get("password")
     db.session.commit()
+
     return jsonify({"message": f"Usuário {id_user} atualizado com sucesso"})
   
   return jsonify({"message": "Usuario não encontrado"}), 404
@@ -75,13 +88,18 @@ def update_user(id_user):
 @login_required
 def delete_user(id_user):
   user = User.query.get(id_user)
+
+  if current_user.role != 'admin':
+    return jsonify({"message": "Operação não permitida"}), 403
+
   if id_user == current_user.id:
     return jsonify({"message": "Deleção não permitida"}), 403
+
   if user:
     db.session.delete(user)
     db.session.commit()
     return jsonify({"message": f"Usuário {id_user} deletado com sucesso"})
-
+  
   return jsonify({"message": "Usuario não encontrado"}), 404
 
 if __name__ == '__main__':
